PPS-11-refactor: move utils to folder
