+++ src/utils/checkExistGame.ts
@@ -1,0 +1,7 @@
+import { Game } from '../interfaces';
+
+export const checkExistGame = (games: Game[], gameId: string) => {
+  const existGame = games.find((game) => game.id === gameId);
+
+  return existGame ? true : false;
+};
